+++ Ez2Buy.Services/DTOs/Product/ProductDetailDto.cs
@@ -11,8 +11,8 @@ namespace Ez2Buy.Services.DTOs.Product
 		public int Id { get; set; }
 		public string Name { get; set; }
 		public string Description { get; set; }
-		public double ListPrice { get; set; }
-		public double Price { get; set; }
+		public decimal ListPrice { get; set; }
+		public decimal Price { get; set; }
 		public string ImageUrl { get; set; }
 		public int CategoryId { get; set; }
